refactor: split api_types.rs to ir/*.rs
