+++ frb_codegen/src/generator_rust.rs
@@ -1,8 +1,8 @@
 use std::borrow::Cow;
 use std::collections::HashSet;
 
-use crate::api_types::ApiType::*;
-use crate::api_types::*;
+use crate::ir::ApiType::*;
+use crate::ir::*;
 use crate::others::*;
 
 pub const HANDLER_NAME: &str = "FLUTTER_RUST_BRIDGE_HANDLER";
