+++ frb_codegen/src/ir/mod.rs
@@ -1,0 +1,29 @@
+mod comment;
+mod field;
+mod file;
+mod func;
+mod ident;
+mod ty;
+mod ty_boxed;
+mod ty_delegate;
+mod ty_enum;
+mod ty_general_list;
+mod ty_optional;
+mod ty_primitive;
+mod ty_primitive_list;
+mod ty_struct;
+
+pub use comment::*;
+pub use field::*;
+pub use file::*;
+pub use func::*;
+pub use ident::*;
+pub use ty::*;
+pub use ty_boxed::*;
+pub use ty_delegate::*;
+pub use ty_enum::*;
+pub use ty_general_list::*;
+pub use ty_optional::*;
+pub use ty_primitive::*;
+pub use ty_primitive_list::*;
+pub use ty_struct::*;
